modified:   MRtools.py 	modified:   resultReport.py 	modified:   run_pyMatch.py

+++ MRtools.py
@@ -911,8 +911,8 @@ class Match:
                 perVoxTmp[com.name] = 0
                 print comname + " does not have shared voxels with template."
             else:
-                perVoxMap[com.name] = (voxel_in_roi/np.count_nonzero(data)) # Overlap voxels as percentage of component map
-                perVoxTmp[com.name] = (voxel_in_roi/len(coordsRCP))         # Overlap voxels as percentage of template voxels
+                perVoxMap[com.name] = float(voxel_in_roi/np.count_nonzero(data)) # Overlap voxels as percentage of component map
+                perVoxTmp[com.name] = float(voxel_in_roi/len(coordsRCP))         # Overlap voxels as percentage of template voxels
                 print comname + " overlap voxels as percentage of component map is " + str(perVoxMap[com.name])
                 print comname + " overlap voxels as percentage of template image is " + str(perVoxTmp[com.name]) + "\n"
         return perVoxMap,perVoxTmp
